# swagger/swagger.yaml
...
openapi: 3.0.0
info:
  version: 1.1.0
  title: Social Media Api
  description: beng beng
servers:
- url: http://localhost:3001
  description: Local server
paths:
...
  /api/users:
    post:
      tags:
        - Users
      summary: Register a user
      requestBody:
        content:
          application/json:
            schema:
              required:
                - email
                - password
                - first_name
                - last_name
              properties:
                first_name:
                  type: string
                last_name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                date_of_birth:
                  type: string
                bio:
                  type: string
      responses:
        '201': 
          description: User registered
        '400': 
          $ref: '#/components/responses/BadRequestError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /api/users/reset:
    post:
      tags:
        - Users
      summary: Request a password reset
      requestBody:
        content:
          application/json:
            schema:
              required:
                - email
              properties:
                email:
                  type: string
      responses:
        '200': 
          description: Token sent
        '404': 
          $ref: '#/components/responses/NotFoundError'
  /api/users/{_id}:
    get:
      tags: 
        - Users
      summary: Returns single user
      parameters:
        - name: _id
          in: path
          description: User ID
          required: true
          schema:
            type: string
      responses:
        '200': 
          description: A User object
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/User'
        '404': 
            $ref: '#/components/responses/NotFoundError'
  /api/users/{_id}/posts:
    get:
      tags: 
        - Users
      summary: Returns posts of a single user
      parameters:
        - name: _id
          in: path
          description: User ID
          required: true
          schema:
            type: string
      responses:
        '200': 
          description: Array of Post objects
          content:
            application/json:
              schema: 
                type: array
                items:  
                  $ref: '#/components/schemas/Post'
        '404': 
            $ref: '#/components/responses/NotFoundError'
  /api/users/{_id}/photos:
    get:
      tags: 
        - Users
      summary: Returns photos of a single user
      parameters:
        - name: _id
          in: path
          description: User ID
          required: true
          schema:
            type: string
      responses:
        '200': 
          description: Array of Photo objects
          content:
            application/json:
              schema: 
                type: array
                items:  
                  $ref: '#/components/schemas/Photo'
        '404': 
            $ref: '#/components/responses/NotFoundError'
  /api/users/{_id}/follow:
    post:
      tags: 
        - Users
      summary: Follow or Unfollow a user
      description: This request will follow the user if it wasn't already followed or unfollow if it was already followed
      parameters:
        - name: _id
          in: path
          description: User ID
          required: true
          schema:
            type: string
      responses:
        '200': 
          description: Unfollwed user
        '201': 
          description: Followed user. Returns a Follow object
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/Follow'
        '404': 
            $ref: '#/components/responses/NotFoundError'
  /api/users/{_id}/follows:
    get:
      tags: 
        - Users
      summary: Returns list of users following and being followed by user
      parameters:
        - name: _id
          in: path
          description: User ID
          required: true
          schema:
            type: string
      responses:
        '200': 
          content:
              application/json:
                schema: 
                  type: object
                  properties: 
                    followersAmount: 
                      type: number
                    followers:
                      type: array
                      items:  
                        $ref: '#/components/schemas/User'
                    followingAmount: 
                      type: number
                    following:
                      type: array
                      items:  
                        $ref: '#/components/schemas/User'
        '40': 
            $ref: '#/components/responses/BadRequestError'
        '404': 
            $ref: '#/components/responses/NotFoundError'
  /api/users/search:
    get:
      tags: 
        - Users
      summary: Search Users
      parameters:
        - name: search
          in: query
          schema:
            type: string
          required: true
      responses:
        '200': 
          description: Array of User objects
          content:
            application/json:
              schema: 
                type: array
                items: 
                  $ref: '#/components/schemas/User'
  /api/posts/:
    post:
      tags:
        - Posts
      summary: Add a post
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - content
              properties:
                content:
                  type: string
                postImage:
                  type: file
          application/json:
            schema:
              required:
                - content
              properties:
                content:
                  type: string
      responses:
        '201': 
          description: Created Post JSON object
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/posts/{_id}:
    get:
      tags: 
        - Posts
      summary: Returns single Post
      parameters:
        - name: _id
          in: path
          description: Post ID
          required: true
          schema:
            type: string
      responses:
        '200': 
          description: A Post object
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/Post'
        '404': 
            $ref: '#/components/responses/NotFoundError'
    
    put:
      tags: 
        - Posts
      summary: Edit a Post
      parameters:
        - name: _id
          in: path
          description: Post ID
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              required:
                - content
              properties:
                content:
                  type: string
      responses:
        '200': 
          description: A Post object
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/Post'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404': 
            $ref: '#/components/responses/NotFoundError'

    delete:
      tags: 
        - Posts
      summary: Delete a Post
      parameters:
        - name: _id
          in: path
          description: Post ID
          required: true
          schema:
            type: string
      responses:
        '200': 
          description: Post Removed
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404': 
            $ref: '#/components/responses/NotFoundError'
...
  /api/comments/{_id}:
    get:
      tags: 
        - Comments
      summary: Returns single Comment
      parameters:
        - name: _id
          in: path
          description: Comment ID
          required: true
          schema:
            type: string
      responses:
        '200': 
          description: A Comment object
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/Comment'
        '404': 
            $ref: '#/components/responses/NotFoundError'
    put:
      tags: 
        - Comments
      summary: Edit a Comment
      parameters:
        - name: _id
          in: path
          description: Comment ID
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              required:
                - content
              properties:
                content:
                  type: string
      responses:
        '200': 
          description: A Comment object
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/Comment'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404': 
            $ref: '#/components/responses/NotFoundError'
    delete:
      tags: 
        - Comments
      summary: Delete a Comment
      parameters:
        - name: _id
          in: path
          description: Comment ID
          required: true
          schema:
            type: string
      responses:
        '200': 
          description: Comment Removed
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404': 
            $ref: '#/components/responses/NotFoundError'
...
components:
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          example: 5da0ff5a9d4a231ad0c9996a
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
        bio:
          type: string
        create_date:
          type: date
          example: "2019-10-01T18:44:43.886Z"
        posts:
          type: array
          items:
            $ref: '#/components/schemas/Post'
        photos:
          type: array
          items:
            $ref: '#/components/schemas/Photo'
        profilePic:
          type: string
        activated:
          type: boolean
        date_of_birth:
          type: date
        isFollowed:
          type: boolean
...
    Post:
      type: object
      properties:
        _id:
          type: string
          example: 5da0ff5a9d4a231ad0c9996a
        author:
          $ref: '#/components/schemas/User'
        create_date:
          type: date
          example: "2019-10-01T18:44:43.886Z"
        last_edit_date:
          type: date
          example: "2019-10-01T18:44:43.886Z"
        content:
          type: string
        photo:
          $ref: '#/components/schemas/Photo'
        edit_history:
          type: array
          items:
            type: object
            properties:
              content:
                type: string
              time:
                type: string
                example: "2019-10-01T18:44:43.886Z"
        likes:
          type: array
          items:
            $ref: '#/components/schemas/User'
        likesAmount:
          type: number
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        commentsAmount:
          type: number
        followers:
          type: array
          items:
            $ref: '#/components/schemas/User'
    Comment:
      type: object
      properties:
        _id:
          type: string
          example: 5da0ff5a9d4a231ad0c9996a
        author:
          $ref: '#/components/schemas/User'
        create_date:
          type: date
          example: "2019-10-01T18:44:43.886Z"
        last_edit_date:
          type: date
          example: "2019-10-01T18:44:43.886Z"
        content:
          type: string
        edit_history:
          type: array
          items:
            type: object
            properties:
              content:
                type: string
              time:
                type: string
                example: "2019-10-01T18:44:43.886Z"
        likes:
          type: array
          items:
            $ref: '#/components/schemas/User'
        likesAmount:
          type: number
        parentPost:
          $ref: '#/components/schemas/Post'
    Photo:
      type: object
      properties:
        _id:
          type: string
          example: 5da0ff5a9d4a231ad0c9996a
        author:
          $ref: '#/components/schemas/User'
        create_date:
          type: date
          example: "2019-10-01T18:44:43.886Z"
        url:
          type: string
        parentPost:
          $ref: '#/components/schemas/Post'
    Follow:
      type: object
      properties:
        follower:
          $ref: '#/components/schemas/User'
        followee:
          $ref: '#/components/schemas/User'
...
  securitySchemes:
    JWT token:            
      type: http
      scheme: bearer
      bearerFormat: JWT 
  responses:
...
    BadRequestError: #400
      description: Bad request format or required parameter is missing
    ForbiddenError: #401
      description: User is not allowed to have access to this resource
    UnauthorizedError: #403
      description: Access token is missing or invalid
    NotFoundError: #404
      description: Item not found   
    ConflictError: #409
      description: Resource already exists
security:
  - JWT token: []